+++ ChangeLog.md
@@ -4,7 +4,7 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v2.13.3 -- not yet released
+## vX.Y.Z -- YYYY-MM-DD
 
 ### Changed
 * Provide support via GitHub discussions.